Moved capi multiple and image code into models.
