+++ commercial/app/controllers/commercial/ContentApiOffersController.scala
@@ -14,10 +14,6 @@ import scala.concurrent.Future
 import scala.concurrent.duration._
 import scala.util.control.NonFatal
 
-import views.support.ImgSrc
-import cards.{Half, Third, Standard}
-import layout.{FaciaWidths, ItemClasses}
-
 sealed abstract class SponsorType(val className: String)
 case object PaidFor extends SponsorType("paidfor")
 case object Supported extends SponsorType("supported")
@@ -123,53 +119,6 @@ class ContentApiOffersController(contentApiClient: ContentApiClient, capiAgent: 
 
   // ----- cAPI Native Templates ----- //
 
-  // Puts together image source info using data from cAPI.
-  private def buildImageData(imageData: Option[ImageMedia]): ImageInfo = {
-
-    val fallbackImageUrl = imageData flatMap ImgSrc.getFallbackUrl
-    val cardType = Standard
-
-    val breakpointWidths = FaciaWidths.mediaFromItemClasses(ItemClasses(
-      mobile = Standard,
-      tablet = cardType,
-      desktop = Some(cardType)
-    )).breakpoints
-
-    val sources = breakpointWidths.map { breakpointWidth =>
-      ImageSource(
-        breakpointWidth.breakpoint.minWidth.getOrElse("0").toString,
-        breakpointWidth.width.toString,
-        ImgSrc.srcsetForBreakpoint(breakpointWidth, breakpointWidths, None,
-          imageData, hidpi = true),
-        ImgSrc.srcsetForBreakpoint(breakpointWidth, breakpointWidths, None,
-          imageData)
-      )
-    }
-
-    ImageInfo(sources, fallbackImageUrl.getOrElse(""))
-
-  }
-
-  // Holds the source element data for the images.
-  case class ImageSource (
-    minWidth: String,
-    sizes: String,
-    hidpiSrcset: String,
-    lodpiSrcset: String
-  )
-
-  object ImageSource {
-    implicit val writesImageSource: Writes[ImageSource] =
-      Json.writes[ImageSource]
-  }
-
-  // Holds all source element data, and the backup image src for older browsers.
-  case class ImageInfo (sources: Seq[ImageSource], backupSrc: String)
-
-  object ImageInfo {
-    implicit val writesImageInfo: Writes[ImageInfo] = Json.writes[ImageInfo]
-  }
-
   // The information needed to render the native cAPI single ad.
   case class CapiSingle(
     articleHeadline: String,
@@ -201,37 +150,6 @@ class ContentApiOffersController(contentApiClient: ContentApiClient, capiAgent: 
     }
 
     implicit val writesCapiSingle: Writes[CapiSingle] = Json.writes[CapiSingle]
-  }
-
-  // The information needed to render the native cAPI multiple ad.
-  case class CapiMultiple(articles: Seq[CapiSingle])
-
-  object CapiMultiple {
-
-    def fromContent(articles: Seq[ContentType]): CapiMultiple = {      
-
-      CapiMultiple(articles.map { article => 
-
-        val content = article.content
-        val imageInfo = buildImageData(content.trail.trailPicture)
-
-        CapiSingle(
-          content.trail.headline,
-          content.metadata.webUrl,
-          content.trail.fields.trailText,
-          imageInfo,
-          content.tags.isAudio,
-          content.tags.isGallery,
-          content.tags.isVideo
-        )
-
-      })
-
-    }
-
-    implicit val writesCapiMultiple: Writes[CapiMultiple] =
-      Json.writes[CapiMultiple]
-
   }
 
   private def renderNative(format: Format, isMulti: Boolean) = Action.async { implicit request =>
